requirements.txt
Flask==1.1.2
gunicorn==20.1.0
pymongo==3.11.4
dnspython==2.0.0
flask-login==0.4.1
flask_socketio==5.0.1
Werkzeug==1.0.1
Pillow==7.0.0
